Fix: refactor iocli project and structure
