+++ iocli/gen/inject/rpc_client_test.go
